Remove LoadRuleSetsTask.
I see preloading RuleSets doesn't provide any tasks execution speed benifit.
