+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/DisabledRulesTest.kt
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 import org.jlleitschuh.gradle.ktlint.KtlintBasePlugin.Companion.LOWEST_SUPPORTED_GRADLE_VERSION
-import org.jlleitschuh.gradle.ktlint.tasks.LoadRuleSetsTask
+import org.jlleitschuh.gradle.ktlint.tasks.KtLintCheckTask
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.File
@@ -117,7 +117,9 @@ abstract class DisabledRulesTest : AbstractPluginTest() {
         projectRoot.withCleanSources()
 
         buildAndFail(CHECK_PARENT_TASK_NAME).apply {
-            assertThat(task(":${LoadRuleSetsTask.TASK_NAME}")?.outcome).isEqualTo(TaskOutcome.FAILED)
+            assertThat(
+                task(":${KtLintCheckTask.buildTaskNameForSourceSet("main")}")?.outcome
+            ).isEqualTo(TaskOutcome.FAILED)
             assertThat(output).contains("Rules disabling is supported since 0.34.2 ktlint version.")
         }
     }
